requirements.txt
pyMUPDF
python-docx
streamlit
pinecone
sentence-transformers
ollama